refactor(publisher): move github publish logic to own file
set up for theoretical multiple publishers
